Change ignore functionality; move tree format to separate flag
Previously ignores where taken from a file, now they are taken from straight from the flag (analog to seek).

it used to have tree format as the default. now it is enabled in
a separate flag. new default prints whole paths to files and direcories.
the reason for this flag was to use this program in vim, similar to
seek. however, i discovered, that vim's command 'gf' can jump from tree
formatted output as well.